moodai/requirements.txt
flask
flask-socketio
tensorflow
opencv-python
pillow
numpy
python-engineio
python-socketio
eventlet
gunicorn